requirements.txt
# ═══════════════════════════════════════════
# TG-Hub V2: Зависимости
# ═══════════════════════════════════════════

# Telegram Bot (aiogram 3.x с поддержкой Business API)
aiogram>=3.4.0

# Web server для webhooks
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# HTTP клиент
aiohttp>=3.9.0
httpx>=0.26.0

# Конфигурация
python-dotenv>=1.0.0

# AI: Claude API
anthropic>=0.18.0

# AI: OpenAI Embeddings
openai>=1.12.0

# Vector Store (RAG)
lancedb>=0.5.0
pyarrow>=15.0.0

# Документы для RAG
pypdf>=4.0.0
python-docx>=1.1.0

# Утилиты
pyyaml>=6.0.0
